add get and placeholder from post
