+++ app.spec.js
@@ -1,0 +1,16 @@
+import request from 'supertest';
+import '@babel/polyfill';
+import app from './server s';
+
+
+
+describe('App', () => {
+  let notes
+  beforeEach(() => {
+    notes = [
+      {
+      id: 1, title: 'take out'
+    }
+  ]
+  })
+})
