+++ package.json
@@ -4,15 +4,22 @@
   "description": "",
   "main": "server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node server.js",
-    "lint": "eslint src/"
+    "lint": "eslint src/",
+    "test": "jest --watch"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
     "express": "^4.17.0",
-    "short-id": "^0.1.0"
+    "shortid": "^2.2.14"
+  },
+  "devDependencies": {
+    "@babel/polyfill": "^7.4.4",
+    "babel-jest": "^24.8.0",
+    "jest": "^24.8.0",
+    "nodemon": "^1.19.0",
+    "supertest": "^4.0.2"
   }
 }
