+++ server s.js
@@ -1,0 +1,9 @@
+import app from './app';
+app.set('port', process.env.PORT || 3000)
+
+
+
+app.listen(app.get('port'), () => {
+  console.log(`${app.locals.notes} is running on http://localhost:${app.get('port')}.`);
+});
+
